код исправлен в соответствии с замечаниями
