+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -5,54 +5,37 @@ import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.storage.FilmStorage;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Map;
 
 @RestController
 @RequestMapping("/films")
 public class FilmController {
-
-    static final int MAX_LENGTH_DESCRIPTION = 200;
-    LocalDate checkData = LocalDate.of(1895, 12, 28);
-    private final Map<Integer, Film> films = new HashMap<>();
+    private static final int MAX_LENGTH_DESCRIPTION = 200;
+    static LocalDate checkData = LocalDate.of(1895, 12, 28);
     private static final Logger log = LoggerFactory.getLogger(FilmController.class);
-
-    protected int idGeneratorForFilm = 0;
-
-    private int generateIdForFilm() {
-        return ++idGeneratorForFilm;
-    }
+    FilmStorage filmStorage = new FilmStorage();
 
     @PostMapping
     public Film create(@RequestBody Film film) {
-        film.setId(generateIdForFilm());
-            validateFilm(film);
-            films.put(film.getId(), film);
-            return film;
+        validateFilm(film);
+        return filmStorage.create(film);
     }
 
     @PutMapping
     public Film put(@RequestBody Film film) {
         validateFilm(film);
-        if (films.containsKey(film.getId())) {
-            films.put(film.getId(), film);
-        } else {
-            log.debug("Попытка обновить отсуствующий в Map<Integer, Film> films фильм");
-            throw new ValidationException("Фильм, который вы хотите обновить, отсутствует в вашей коллекции");
-        }
-        return film;
+        return filmStorage.put(film);
     }
 
     @GetMapping
     public ArrayList<Film> findAll() {
-        log.debug("Текущее количество фильмоф: {}", films.size());
-        return new ArrayList<>(films.values());
+        return filmStorage.findAll();
     }
 
-    protected void validateFilm(@RequestBody Film film) {
+    private static void validateFilm(@RequestBody Film film) {
         if (film.getName() == null || film.getName().isEmpty()) {
             log.debug("Отсутствует название фильма");
             throw new ValidationException("Не указано название фильма");
