+++ src/main/java/ru/yandex/practicum/filmorate/exception/UserAlreadyExistException.java
@@ -1,6 +1,6 @@
 package ru.yandex.practicum.filmorate.exception;
 
-public class UserAlreadyExistException extends Throwable {
+public class UserAlreadyExistException extends RuntimeException {
     public UserAlreadyExistException(final String message) {
         super(message);
     }
